requirements.txt
python-dotenv
langchain-anthropic
langchain-openai
langchain-core
magika
streamlit
langchain
tiktoken
anthropic==0.37.1